+++ src/app/common/index.ts
@@ -2,3 +2,4 @@ export * from './ellipsis-title.directive';
 export * from './window-resize-tracker.service';
 export * from './not-found';
 export * from './logging';
+export * from './types';
