+++ src/app/common/logging/prefixed-console-logger-factory.service.ts
@@ -2,7 +2,7 @@ import { Inject, Injectable, Optional } from '@angular/core';
 import { ILogger } from './i-logger';
 import { PrefixedConsoleLogger } from './prefixed-console-logger';
 import { LOG_LEVEL, LogLevel } from './log-levels';
-import { ExtractTokenType } from '../../types';
+import { ExtractTokenType } from '../types';
 
 @Injectable({ providedIn: 'root' })
 export class PrefixedConsoleLoggerFactoryService {
