+++ src/app/common/types/browser-api.ts
