+++ src/app/common/types/type-helpers.ts
