+++ src/app/control-schemes/view/control-scheme-view-io/control-scheme-view-io.component.ts
@@ -4,7 +4,7 @@ import { JsonPipe, NgIf } from '@angular/common';
 import { MatCardModule } from '@angular/material/card';
 import { TranslocoModule } from '@ngneat/transloco';
 import { MatIconModule } from '@angular/material/icon';
-import { PortCommandTask } from '../../../types';
+import { PortCommandTask } from '../../../common';
 import { PORT_TASK_TYPE_TO_L10N } from '../../../i18n';
 
 type LastExecutedTaskTemplateDate = {
