+++ src/app/i18n/port-task-type-to-l10n.ts
@@ -1,4 +1,4 @@
-import { PortCommandTaskType } from '../types';
+import { PortCommandTaskType } from '../common';
 
 export const PORT_TASK_TYPE_TO_L10N: { [k in PortCommandTaskType]: string } = {
     [PortCommandTaskType.SetSpeed]: 'portCommandTaskTypeSetLinearSpeed',
