+++ src/app/store/actions/hub-port-tasks.actions.ts
@@ -1,5 +1,5 @@
 import { createActionGroup, props } from '@ngrx/store';
-import { PortCommandTask } from '../../types';
+import { PortCommandTask } from '../../common';
 
 export const HUB_PORT_TASKS_ACTIONS = createActionGroup({
     source: 'Hub Port Tasks',
