+++ src/app/store/entity-adapters.ts
@@ -12,7 +12,7 @@ import {
     PortModeInfo
 } from './i-state';
 import { IOType } from '../lego-hub';
-import { PortCommandTask } from '../types';
+import { PortCommandTask } from '../common';
 
 export const HUB_ATTACHED_IOS_ENTITY_ADAPTER: EntityAdapter<AttachedIO> = createEntityAdapter<AttachedIO>({
     selectId: (io) => hubAttachedIosIdFn(io.hubId, io.portId),
