+++ src/app/tasks-processing/task-composer/port-command-task-composer.ts
@@ -1,6 +1,6 @@
 import { IPortCommandTaskComposer } from './i-port-command-task-composer';
 import { ControlSchemeBinding } from '../../store';
-import { PortCommandTask } from '../../types/port-command-task';
+import { PortCommandTask } from '../../common';
 
 export abstract class PortCommandTaskComposer implements IPortCommandTaskComposer {
     private next?: PortCommandTaskComposer;
