+++ src/app/tasks-processing/task-executor/i-task-executor.ts
@@ -1,5 +1,5 @@
 import { Hub } from '../../lego-hub';
-import { PortCommandTask } from '../../types';
+import { PortCommandTask } from '../../common';
 
 export interface ITaskExecutor {
     executeTask(task: PortCommandTask, hub: Hub): Promise<void>;
