+++ src/app/tasks-processing/task-executor/task-executor.ts
@@ -1,6 +1,6 @@
 import { ITaskExecutor } from './i-task-executor';
 import { Hub } from '../../lego-hub';
-import { PortCommandTask } from '../../types';
+import { PortCommandTask } from '../../common';
 
 export abstract class TaskExecutor implements ITaskExecutor {
     private next?: TaskExecutor;
