+++ src/app/tasks-processing/task-queue-compressor/compressors/set-speed-task-queue-compressor.ts
@@ -1,5 +1,5 @@
 import { ITaskSpecificQueueCompressor } from '../i-task-specific-queue-compressor';
-import { PortCommandSetLinearSpeedTask, PortCommandTask, PortCommandTaskType } from '../../../types';
+import { PortCommandSetLinearSpeedTask, PortCommandTask, PortCommandTaskType } from '../../../common';
 
 export class SetSpeedTaskQueueCompressor implements ITaskSpecificQueueCompressor {
     public compress(queue: PortCommandTask[]): PortCommandTask[] {
