+++ src/app/tasks-processing/task-queue-compressor/i-task-specific-queue-compressor.ts
@@ -1,4 +1,4 @@
-import { PortCommandTask } from '../../types';
+import { PortCommandTask } from '../../common';
 
 export interface ITaskSpecificQueueCompressor {
     compress(queue: PortCommandTask[]): PortCommandTask[];
