+++ src/app/tasks-processing/task-suppressor/i-task-suppressor.ts
@@ -1,4 +1,4 @@
-import { PortCommandTask } from '../../types';
+import { PortCommandTask } from '../../common';
 
 export interface ITaskSuppressor {
     shouldSuppressTask<T extends PortCommandTask>(task: T, lastTaskOfKindInQueue: T): boolean;
