+++ src/app/tasks-processing/task-suppressor/suppressors/set-speed-task-suppressor.ts
@@ -1,6 +1,6 @@
 import { TaskSuppressor } from '../task-suppressor';
 import { MotorFeature } from '../../../lego-hub/features';
-import { PortCommandTask, PortCommandTaskType } from '../../../types';
+import { PortCommandTask, PortCommandTaskType } from '../../../common';
 
 export class SetSpeedTaskSuppressor extends TaskSuppressor {
     public static readonly speedThreshold = 10;
